software/client/pyproject.toml
[project]
name = "client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.3",
    "lap>=0.5.12",
    "numpy>=2.4.2",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.0",
    "pydantic-to-typescript>=2.0.0",
    "pyright>=1.1.408",
    "pyserial>=3.5",
    "python-dotenv>=1.2.1",
    "readchar>=4.2.1",
    "requests>=2.32.5",
    "requests-oauthlib>=2.0.0",
    "ultralytics>=8.4.12",
    "uvicorn>=0.40.0",
    "websockets>=16.0",
    "wsproto>=1.3.2",
]

[tool.pyright]
reportMissingImports = "none"
reportMissingModuleSource = "none"
